% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicKineticModel.R
\name{basicKineticModel}
\alias{basicKineticModel}
\title{Basic Kinetic Model Constructor}
\usage{
basicKineticModel(times = NA_real_, synthRate = NA_real_,
  degRate = NA_real_, initAbund = NA_real_, ids = NA, data = NULL,
  expMetadata = data.frame(), dispersionModel = function(x) { })
}
\arguments{
\item{synthRate}{Synthesis rates. Must be greater than 0.}

\item{degRate}{Degredation rate. Must be between 0 and 1.}

\item{initAbund}{Initial abundance at time 0.}

\item{ids}{A character string id.}

\item{data}{Experimental or simulated matrix of count data}

\item{expMetadata}{A data.frame where each row corresponds to the column in data with the same index and the columns are different properties}
}
\description{
Contructs a simple kinetic model object from synthesis and degredation rates
}
