% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateAbundance.R
\name{calculateAbundance}
\alias{calculateAbundance}
\title{Calculate transcript abundance}
\arguments{
\item{object}{A simpleKineticSim object}

\item{times}{Times to calculate abundance values at.}
}
\description{
Calculates transcript abundance
}
\examples{
ts=simpleKineticSim(syn.rate = 1:10,deg.rate = rep(0.3,10))
ts=calculateAbundance(ts,0:30)
}
