% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePosteriors.R
\docType{methods}
\name{calculatePosteriors}
\alias{calculatePosteriors}
\title{Calculate Posterior Probabilities for Infered Parameters}
\usage{
\S4method{calculatePosteriors}{basicKineticModel}(object,
  alphaRange = numeric(2), betaRange = numeric(2), paramSpaceSize = 10^4,
  dispByGene = T, logProbAlpha = NULL, logProbBeta = NULL)
}
\arguments{
\item{object}{A \linkS4class{basicKineticModel} object}

\item{alphaRange}{Scale factors used to calculate the upper and lower bounds of the parameter range explored for \code{alpha}.  These scale factors will
be applied to the infered value of \code{alpha}.  Must be defined as \code{c(lower,upper)}.}

\item{betaRange}{Scale factors used to calculate the upper and lower bounds of the parameter range explored for \code{beta}.  These scale factors will
be applied to the infered value of \code{beta}.  Must be defined as \code{c(lower,upper)}.}

\item{paramSpaceSize}{The total size of parameter space to numerically integrate over. Half of the parameter space will be given to each parameter.}

\item{dispByGene}{Boolean controlling the expected nature of the \code{dispersionModel}. See \code{dispersionModel} description in \code{\link{inferParameters}}
for more details.}

\item{logProbAlpha}{A function that returns the log probability for a given value of \code{alpha}}

\item{lobProbBeta}{A function that returns the log probability for a given value of \code{beta}}
}
\description{
Uses numeric methods to estimate posteriors for the infered parameters \code{alpha} (synthesis rate) and \code{beta} (degredation rate).
Currently uses a flat prior for both alpha and beta as a default. This function is written such that non-flat priors may be used in the future.
}
\examples{
EXAMPLE HERE
}
