% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAbundances.R
\name{plotAbundances}
\alias{plotAbundances}
\title{Plot transcript abundance}
\arguments{
\item{object}{A simpleKineticSim object}
}
\description{
Plots transcript abundance. Must run calculateAboundance first.
}
\examples{
ts=simpleKineticSim(syn.rate = 1:10,deg.rate = rep(0.3,10))
ts=calculateAbundance(ts,0:30)
plotAbundances(ts)
}
