% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMarginalDist.R
\name{plotMarginalDist}
\alias{plotMarginalDist}
\title{Plot the marginal posterior distribution of an estimated parameter}
\usage{
plotMarginalDist(object, geneIdx, param)
}
\arguments{
\item{object}{A \linkS4class{basicKineticModel} object}

\item{geneIdx}{The row index in \code{@data} of the gene from which you want to plot the marginal posterior distribution from}

\item{param}{A string of the name of the parameter you want to plot.}
}
\description{
Plots the marginal distribution of a selected parameter for a given gene
}
