% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPosteriors.R
\name{plotPosteriors}
\alias{plotPosteriors}
\title{Plot Posteriors for Infered Parameters}
\usage{
plotPosteriors(object, geneIdx, alphaRange=NULL,
  betaRange=NULL, relative=T, recalculate=F, ...)
}
\arguments{
\item{object}{A \linkS4class{basicKineticModel} object}

\item{geneIdx}{The gene for which to plot the posterior distribution}

\item{...}{Additional arguments needed for recalculating the posterior distribution. Only used if \code{recalculate = TRUE}. See \code{\link{calculatePosteriors}}
for details on which parameters need to be supplied.}

\item{alphaRange}{Determines the range of alpha values to plot over. Must be specified as \code{c(lower,upper)}. See \code{relative} for more information on how to specify the range. If not specified,
will default to the original range used when \code{\link{calculatePosteriors}} was called.}

\item{betaRange}{Determines the range of beta values to plot over. Must be specified as \code{c(lower,upper)}. See \code{relative} for more information on how to specify the range. If not specified,
will default to the original range used when \code{\link{calculatePosteriors}} was called.}

\item{relative}{Boolean used to determine whether parameter ranges are specified as relative or absolute. If \code{TRUE}, the ranges are treated as multiples of the
estimated parameter to plot over (eg. \code{alphaRange = c(.25,2)} will plot from .25*\code{alpha} to 2*\code{alpha}).  If \code{FALSE} then the ranges are treated
as absolute values of the specified parameter.}

\item{recalculate}{Boolean used to determine if the posterior distribution should be recalculated for the selected gene and parameter boundaries.}
}
\value{
A \code{ggplot2 plot} object containing the posterior plot
}
\description{
Plots the results of the calculations from \code{\link{calculatePosteriors}} for a single gene. Allows for the selection of custom plotting bounds as
well as the recalculation of the posterior distribution within those bounds.

Recalculation of the posterior is useful for obtaining better resolution for
plots that are significantly 'zoomed in' compared to the original posterior calculation. This also allows you to avoid having to call \code{\link{calculatePosteriors}}
if you are only interested in looking at one gene at a time.

Note: \code{\link{calculatePosteriors}} must be run before calling this method unless \code{recalculate = T} in which case it \strong{does not} need to be run.
}
\examples{
##setup
bkm=basicKineticModel(times=0:30,synthRate=1:10,degRate = rep(0.5,10))
bkm=simulateData(bkm)
bkm=simulateReads(bkm,expectedLibSize=10^6,replicates=3,spikeInSizes = 200,dispersionModel=function(x){rep(10^3,length(x))}) #CLUGE
bkm=inferParameters(bkm)
bkm=calculatePosteriors(bkm,alphaRange=c(.25,2))

##plot entire distribution
plotPosteriors(bkm,geneIdx=3)

##plot absolute subset of distribution
plotPosteriors(object=bkm,geneIdx=3,alphaRange=c(3000,4500),betaRange=c(.45,.55),relative=F)

##plot relative subset of distribution
plotPosteriors(object=bkm,geneIdx=3,alphaRange=c(.9,1.1))

##plot subset and recalculate posterior distribution (to obtain higher resolution)
#note: calculatePosteriors does NOT need to be run prior to plotting if recalculate=T
plotPosteriors(object=bkm,geneIdx=3,alphaRange=c(3000,4500),betaRange=c(.45,.55),relative=F,recalculate=T,paramSpaceSize=10^5)


}
