% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictAbundance.R
\name{predictAbundance}
\alias{predictAbundance}
\title{Calculate transcript abundance}
\arguments{
\item{object}{A basicKineticModel object}

\item{times}{Times to calculate abundance values at.}
}
\description{
Calculates transcript abundance
}
\examples{
ts=basicKineticModel(synthRate = 1:10,degRate = rep(0.3,10))
ts=predictAbundance(ts,0:30)
}
