% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-simpleKineticExperiment.R
\name{simpleKineticExperiment-class}
\alias{simpleKineticExperiment-class}
\title{Class simpleKineticExperiment}
\description{
Class \code{simpleKineticExperiment} defines a kinetic model dx/dt=a-b*x
}
