% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleKineticSim.R
\name{simpleKineticSim}
\alias{simpleKineticSim}
\title{Build simulation object}
\usage{
simpleKineticSim(syn.rate, deg.rate, init.abundance = NA, ids = NA)
}
\arguments{
\item{syn.rate}{Synthesis rates. Must be greater than 0.}

\item{deg.rate}{Degredation rate. Must be between 0 and 1.}

\item{init.abundance}{Initial abundance at time 0.}

\item{ids}{A character string id.}
}
\description{
Contructs a simple kinetic simulation object from synthesis and degredation rates
}
