% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateReads.R
\docType{methods}
\name{simulateReads}
\alias{simulateReads}
\title{Simulate Reads}
\usage{
\S4method{simulateReads}{kineticModel}(object, expectedLibSize = 10^6,
  replicates = 2, numSpikeIns = 4, spikeInSizes = NULL,
  dispersionModel = NULL, dispByGene = T)
}
\arguments{
\item{object}{A \linkS4class{kineticModel} object}

\item{expectedLibSize}{The expected total number of reads per sequencing run/batch. May be supplied as a single number which will be applied uniformly,
or a vector of library sizes where each element corresponds to the library depth at a given time point.}

\item{replicates}{Replicates per condition}

\item{numSpikeIns}{The number of unique spike in transcripts used.}

\item{spikeInSizes}{The expected number of reads for each type of spike in used. May be a single number used for all spike-in transcripts
or an array of abundances for each unique spike-in transcript.}

\item{dispersionModel}{A function that takes in a vector of means and returns a vector of dispersions based on those means for a negative bionomial model}
}
\description{
Simulate reads from a \linkS4class{kineticModel} object. Will simulate data if the input \linkS4class{kineticModel} does not already contain simulated data.
}
\examples{
bkm = basicKineticModel(synthRate = 1:10,degRate = rep(0.3,10))
bkm = simulateData(bkm) #optional
bkm = simulateReads(bkm,expectedLibSize=10^6,replicates=3,numSpikeIns=4,spikeInSizes=200,dispersionModel=function(x){rep(10^4,length(x))})
}
