% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateEquilibrium.R
\name{calculateEquilibrium}
\alias{calculateEquilibrium}
\title{Calculate equilibrium transcript abundance}
\arguments{
\item{object}{A simpleKineticSim object}
}
\description{
Calculates equilibrium transcript abundance
}
\examples{
ts=simpleKineticSim(syn.rate = 1:10,deg.rate = rep(0.3,10))
ts=calculateEquilibrium(ts)
}
