% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateDispersions.R
\docType{methods}
\name{estimateDispersions}
\alias{estimateDispersions}
\title{Estimate Dispersions Using DESeq2}
\usage{
\S4method{estimateDispersions}{basicKineticModel}(object, byGene = T)
}
\arguments{
\item{object}{A \linkS4class{basicKineticModel} object}

\item{byGene}{Boolean controlling fuction return value. See \code{description} or \code{value} for more information}
}
\value{
UPDATE ME:::::::If \code{byGene = TRUE} then a vector containing a single dispersion
estimate for each gene will be returned. If \code{FALSE} then a general function will be returned that gives a dispersion estimate
for a given mean using the composite information from all genes, time points, and replicates.
}
\description{
Estimates the dispersions for each gene using DESeq2. If \code{byGene = TRUE} then a vector containing a single dispersion
estimate for each gene will be returned. If \code{FALSE} then a general function will be returned that gives a dispersion estimate
for a given mean using the composite information from all genes, time points, and replicates.
}
