% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferParameters.R
\docType{methods}
\name{inferParameters}
\alias{inferParameters}
\title{Infer Parameters from Read Data}
\usage{
\S4method{inferParameters}{basicKineticModel}(object, dispersionModel = NULL,
  dispByGene = T)
}
\arguments{
\item{object}{A \linkS4class{basicKineticModel} object}
}
\description{
Infers \code{alpha} (synthesis rate) and \code{beta} (degredation rate) from sequencing read data (stored in the \code{data} slot) using the
basic kinetic model: \code{dx/dt = alpha - beta * x}.  If no read data exists in the provided \linkS4class{basicKineticModel} then simulated read data
will be generated using simulation data stored in \code{@simData} before parameter inference. If no simulated data exists, it will be generated
before simulating read data and inferring parameters.
}
\examples{
bkm=basicKineticModel(synthRate = 1:10,degRate = rep(0.3,10), times=0:30)
bkm=simulateData(bkm) #optional
bkm=simulateReads(bkm)
bkm=inferParameters(bkm)
}
