% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAbundances.R
\name{plotAbundances}
\alias{plotAbundances}
\title{Plot transcript abundance}
\arguments{
\item{object}{A basicKineticModel object}
}
\description{
Plots transcript abundance. Must run calculateAbundance first.
}
\examples{
ts=basicKineticModel(synthRate = 1:10,degRate = rep(0.3,10))
ts=calculateAbundance(ts,0:30)
plotAbundances(ts)
}
