% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotParameterFit.R
\docType{methods}
\name{plotParameterFit}
\alias{plotParameterFit}
\title{Plot Data Against the Fit Parameters}
\usage{
\S4method{plotParameterFit}{basicKineticModel}(object, geneIdx = NULL,
  legend = F, scaleData = T, plotTimes = NULL)
}
\arguments{
\item{object}{A \linkS4class{basicKineticModel} object}

\item{geneIdx}{The row index from \code{@data} of the gene to plot. May be defined as an integer for one gene or a vector for multiple genes.}

\item{legend}{Boolean controlling if the plot should contain a legend.}

\item{scaleData}{If true, will scale the data by a given genes's corresponding normalization factors. If false, will scale the fit function by the average
normalization factor across replicates for each time point.}

\item{plotTimes}{A vector of times specifying the times over which to plot the data and parameter curve. All time points in \code{object@data} that are also in
\code{plotTimes} will be plotted. Time points \strong{only} in \code{plotTimes} will be ignored.  If \code{NULL} the times specified by \code{object@times}
will be used.}
}
\description{
Plot the curve defined by the estimated parameters against the raw data for a selected set of genes
}
