% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleKineticExperiment.R
\name{simpleKineticExperiment}
\alias{simpleKineticExperiment}
\title{Build simulation object}
\usage{
simpleKineticExperiment(syn.rate, deg.rate, init.abundance = NA, ids = NA,
  data = NULL, col.info = data.frame(), design = expression())
}
\arguments{
\item{syn.rate}{Synthesis rates. Must be greater than 0.}

\item{deg.rate}{Degredation rate. Must be between 0 and 1.}

\item{init.abundance}{Initial abundance at time 0.}

\item{ids}{A character string id.}

\item{data}{Experimental or simulated matrix of count data}

\item{col.info}{A data.frame where each row corresponds to the column in data with the same index and the columns are different properties}

\item{design}{An expression that describes the experimental design}
}
\description{
Contructs a simple kinetic simulation object from synthesis and degredation rates
}
