% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateReads.R
\name{simulateReads}
\alias{simulateReads}
\title{Simulate Reads}
\arguments{
\item{object}{A kineticSim object}

\item{expected.lib.size}{The expected library sequencing depth. May vary from this based on sampling.}

\item{replicates}{Replicates per condition}

\item{times}{A vector of times}

\item{errorModel}{A function that takes in a vector of values and returns a vector of dispersions for a negative bionomial model}
}
\description{
Simulate Reads from kineticModel object
}
\examples{
ts=basicKineticModel(synthRate = 1:10,degRate = rep(0.3,10))
ts=calculateEquilibrium(ts)
}
