% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateReads.R
\name{simulateReads}
\alias{simulateReads}
\title{Simulate Reads}
\arguments{
\item{object}{A kineticModel object}

\item{expectedLibSize}{The expected library sequencing depth. May vary from this based on sampling.}

\item{replicates}{Replicates per condition}

\item{times}{A vector of times}

\item{errorModel}{A function that takes in a vector of values and returns a vector of dispersions for a negative bionomial model}
}
\description{
Simulate Reads from kineticModel object
}
\examples{
bkm = basicKineticModel(synthRate = 1:10,degRate = rep(0.3,10))
bkm = simulateData(bkm) #optional
bkm = simulateReads(bkm,expectedLibSize=3,replicates=1,errorModel=function(x){rep(2,length(x))})
}
