% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeEpsilonEq.R
\name{timeEpsilonEq}
\alias{timeEpsilonEq}
\title{Calculate time to near equilibrium}
\arguments{
\item{object}{A basicKineticModel object}

\item{epsilon.percent}{+/- Percent of equilibrium that must be reached}
}
\description{
Calculates the time at which each transcript will be within +/- epsilon% of its equilibrium value
}
\examples{
ts=basicKineticModel(synthRate = 1:10,degRate = rep(0.3,10))
ts=timeEquilibrium(ts,0.01) # To 99\% or 101\% of equilibrium value
}
